+++ config/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.18.0](https://github.com/klarna/gram/compare/v4.17.6...v4.18.0) (2024-11-18)
+
+**Note:** Version bump only for package @gram/config
+
 ## [4.17.6](https://github.com/klarna/gram/compare/v4.17.5...v4.17.6) (2024-10-28)
 
 **Note:** Version bump only for package @gram/config